requirements.txt
django-oscar==0.6.4
-e git+git@github.com:labxio/django-oscar-stripe.git#egg=django-oscar-stripe
South
Whoosh

# Heroku
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
django-storages==1.1.8
boto==2.27.0